+++ src/main/runtime/rpc/methods/ui-state-schema-parity-checks.ts
@@ -1,14 +1,6 @@
 import type { z } from 'zod'
-import type {
-  PersistedUIState,
-  RightSidebarTab,
-  WorktreeCardProperty
-} from '../../../../shared/types'
-import type {
-  RightSidebarTabSchema,
-  UiUpdateFieldsSchema,
-  WorktreeCardPropertySchema
-} from './client-ui-schemas'
+import type { PersistedUIState } from '../../../../shared/types'
+import type { UiUpdateFieldsSchema } from './client-ui-schemas'
 import type { AssertNoMissingKeys, AssertNoMissingValues } from './ui-state-schema-parity'
 
 // Why: state only the main process ever writes (store.updateUI, star-nag's own
@@ -33,13 +25,11 @@ void _uiUpdateParity
 
 // Why: key parity is blind to enum drift, which is how 'cli' and three
 // rightSidebarTab members went missing while the guard above stayed green.
-const _worktreeCardPropertyParity: AssertNoMissingValues<
-  WorktreeCardProperty,
-  z.infer<WorktreeCardPropertySchema>
+// Checked over every shared key, not a hand-picked pair — naming the two known
+// offenders would leave the next field to drift exactly as unguarded.
+// z.input, not z.infer: what a client may SEND, before `.transform()` narrows it.
+const _uiUpdateValueParity: AssertNoMissingValues<
+  Omit<PersistedUIState, MainOwnedUIState>,
+  z.input<UiUpdateFieldsSchema>
 > = true
-void _worktreeCardPropertyParity
-const _rightSidebarTabParity: AssertNoMissingValues<
-  RightSidebarTab,
-  z.infer<RightSidebarTabSchema>
-> = true
-void _rightSidebarTabParity
+void _uiUpdateValueParity
